Pull out sort into its own file
