+++ Sources/TuistGenerator/Generator/ProjectGenerator.swift
@@ -106,16 +106,17 @@ final class ProjectGenerator: ProjectGenerating {
                               archiveVersion: projectConstants.archiveVersion,
                               classes: [:])
 
-        let groups = ProjectGroups.generate(project: project, pbxproj: pbxproj, xcodeprojPath: xcodeprojPath, sourceRootPath: sourceRootPath)
+        let groups = ProjectGroups.generateInitialGroups(project: project,
+                                                         pbxproj: pbxproj,
+                                                         xcodeprojPath: xcodeprojPath,
+                                                         sourceRootPath: sourceRootPath)
+
         let fileElements = ProjectFileElements()
         try fileElements.generateProjectFiles(project: project,
                                               graph: graph,
                                               groups: groups,
                                               pbxproj: pbxproj,
                                               sourceRootPath: sourceRootPath)
-
-        ProjectGroups.sortGroups(group: groups.main)
-        ProjectGroups.addFirstLevelDefaults(firstLevelGroup: groups)
 
         let configurationList = try configGenerator.generateProjectConfig(project: project, pbxproj: pbxproj, fileElements: fileElements)
         let pbxProject = try generatePbxproject(project: project,
@@ -127,7 +128,6 @@ final class ProjectGenerator: ProjectGenerating {
         let nativeTargets = try generateTargets(project: project,
                                                 pbxproj: pbxproj,
                                                 pbxProject: pbxProject,
-                                                groups: groups,
                                                 fileElements: fileElements,
                                                 sourceRootPath: sourceRootPath,
                                                 graph: graph)
@@ -159,7 +159,7 @@ final class ProjectGenerator: ProjectGenerating {
         let pbxProject = PBXProject(name: project.name,
                                     buildConfigurationList: configurationList,
                                     compatibilityVersion: Xcode.Default.compatibilityVersion,
-                                    mainGroup: groups.main,
+                                    mainGroup: groups.sortMainAndAddDefaultGroups(),
                                     developmentRegion: Xcode.Default.developmentRegion,
                                     hasScannedForEncodings: 0,
                                     knownRegions: knownRegions,
@@ -177,7 +177,6 @@ final class ProjectGenerator: ProjectGenerating {
     private func generateTargets(project: Project,
                                  pbxproj: PBXProj,
                                  pbxProject: PBXProject,
-                                 groups _: ProjectGroups,
                                  fileElements: ProjectFileElements,
                                  sourceRootPath: AbsolutePath,
                                  graph: Graphing) throws -> [String: PBXNativeTarget] {
