+++ Tests/TuistGeneratorTests/Generator/ProjectGroupsTests.swift
@@ -42,14 +42,13 @@ final class ProjectGroupsTests: XCTestCase {
                 return []
             }
         }
-        subject = ProjectGroups.generate(project: project,
-                                         pbxproj: pbxproj,
-                                         xcodeprojPath: project.path.appending(component: "\(project.fileName).xcodeproj"),
-                                         sourceRootPath: sourceRootPath,
-                                         playgrounds: playgrounds)
-        ProjectGroups.addFirstLevelDefaults(firstLevelGroup: subject)
+        subject = ProjectGroups.generateInitialGroups(project: project,
+                                                      pbxproj: pbxproj,
+                                                      xcodeprojPath: project.path.appending(component: "\(project.fileName).xcodeproj"),
+                                                      sourceRootPath: sourceRootPath,
+                                                      playgrounds: playgrounds)
 
-        let main = subject.main
+        let main = subject.sortMainAndAddDefaultGroups()
         XCTAssertNil(main.path)
         XCTAssertEqual(main.sourceTree, .group)
         XCTAssertEqual(main.children.count, 5)
@@ -83,11 +82,11 @@ final class ProjectGroupsTests: XCTestCase {
         playgrounds.pathsStub = { _ in
             []
         }
-        subject = ProjectGroups.generate(project: project,
-                                         pbxproj: pbxproj,
-                                         xcodeprojPath: project.path.appending(component: "\(project.fileName).xcodeproj"),
-                                         sourceRootPath: sourceRootPath,
-                                         playgrounds: playgrounds)
+        subject = ProjectGroups.generateInitialGroups(project: project,
+                                                      pbxproj: pbxproj,
+                                                      xcodeprojPath: project.path.appending(component: "\(project.fileName).xcodeproj"),
+                                                      sourceRootPath: sourceRootPath,
+                                                      playgrounds: playgrounds)
 
         XCTAssertNil(subject.playgrounds)
     }
@@ -106,20 +105,19 @@ final class ProjectGroupsTests: XCTestCase {
                                    targets: [target1, target2, target3, target4])
 
         // When
-        subject = ProjectGroups.generate(project: project,
-                                         pbxproj: pbxproj,
-                                         xcodeprojPath: project.path.appending(component: "\(project.fileName).xcodeproj"),
-                                         sourceRootPath: sourceRootPath,
-                                         playgrounds: playgrounds)
-        ProjectGroups.addFirstLevelDefaults(firstLevelGroup: subject)
+        subject = ProjectGroups.generateInitialGroups(project: project,
+                                                      pbxproj: pbxproj,
+                                                      xcodeprojPath: project.path.appending(component: "\(project.fileName).xcodeproj"),
+                                                      sourceRootPath: sourceRootPath,
+                                                      playgrounds: playgrounds)
 
         // Then
-        let paths = subject.main.children.compactMap { $0.nameOrPath }
+        let paths = subject.sortMainAndAddDefaultGroups().children.compactMap { $0.nameOrPath }
         XCTAssertEqual(paths, [
-            "P",
+            "A",
             "B",
             "C",
-            "A",
+            "P",
             "Frameworks",
             "Playgrounds",
             "Products",
@@ -127,11 +125,11 @@ final class ProjectGroupsTests: XCTestCase {
     }
 
     func test_targetFrameworks() throws {
-        subject = ProjectGroups.generate(project: project,
-                                         pbxproj: pbxproj,
-                                         xcodeprojPath: project.path.appending(component: "\(project.fileName).xcodeproj"),
-                                         sourceRootPath: sourceRootPath,
-                                         playgrounds: playgrounds)
+        subject = ProjectGroups.generateInitialGroups(project: project,
+                                                      pbxproj: pbxproj,
+                                                      xcodeprojPath: project.path.appending(component: "\(project.fileName).xcodeproj"),
+                                                      sourceRootPath: sourceRootPath,
+                                                      playgrounds: playgrounds)
 
         let got = try subject.targetFrameworks(target: "Test")
         XCTAssertEqual(got.name, "Test")
@@ -141,11 +139,11 @@ final class ProjectGroupsTests: XCTestCase {
 
     func test_projectGroup_unknownProjectGroups() throws {
         // Given
-        subject = ProjectGroups.generate(project: project,
-                                         pbxproj: pbxproj,
-                                         xcodeprojPath: project.path.appending(component: "\(project.fileName).xcodeproj"),
-                                         sourceRootPath: sourceRootPath,
-                                         playgrounds: playgrounds)
+        subject = ProjectGroups.generateInitialGroups(project: project,
+                                                      pbxproj: pbxproj,
+                                                      xcodeprojPath: project.path.appending(component: "\(project.fileName).xcodeproj"),
+                                                      sourceRootPath: sourceRootPath,
+                                                      playgrounds: playgrounds)
 
         // When / Then
         XCTAssertThrowsSpecific(try subject.projectGroup(named: "abc"),
@@ -160,11 +158,11 @@ final class ProjectGroupsTests: XCTestCase {
         let project = Project.test(filesGroup: .group(name: "P"),
                                    targets: [target1, target2, target3])
 
-        subject = ProjectGroups.generate(project: project,
-                                         pbxproj: pbxproj,
-                                         xcodeprojPath: project.path.appending(component: "\(project.fileName).xcodeproj"),
-                                         sourceRootPath: sourceRootPath,
-                                         playgrounds: playgrounds)
+        subject = ProjectGroups.generateInitialGroups(project: project,
+                                                      pbxproj: pbxproj,
+                                                      xcodeprojPath: project.path.appending(component: "\(project.fileName).xcodeproj"),
+                                                      sourceRootPath: sourceRootPath,
+                                                      playgrounds: playgrounds)
 
         // When / Then
         XCTAssertNotNil(try? subject.projectGroup(named: "A"))
@@ -178,182 +176,5 @@ final class ProjectGroupsTests: XCTestCase {
 
     func test_projectGroupsError_type() {
         XCTAssertEqual(ProjectGroupsError.missingGroup("abc").type, .bug)
-    }
-
-    func test_projectGroupsSort_simpleGroupsCase() throws {
-        // Given
-        let mainGroup = PBXGroup(children: [
-            file("somefile1.swift"),
-            group("somegroup2", [
-                file("somefile2.swift"),
-                file("somefile1.swift"),
-            ]),
-            group("somegroup1", [
-                file("somefile4.swift"),
-                file("somefile3.swift"),
-            ]),
-        ])
-
-        // When
-        ProjectGroups.sortGroups(group: mainGroup)
-
-        // Then
-        assertGroupsEqual(mainGroup, group("project", [
-            group("somegroup1", [
-                file("somefile3.swift"),
-                file("somefile4.swift"),
-            ]),
-            group("somegroup2", [
-                file("somefile1.swift"),
-                file("somefile2.swift"),
-            ]),
-            file("somefile1.swift"),
-        ]))
-    }
-
-    func test_projectGroupsSort_nestedGroupsCase() throws {
-        // Given
-        let mainGroup = PBXGroup(children: [
-            file("somefile1.swift"),
-            group("somegroup2", [
-                file("somefile2.swift"),
-                file("somefile1.swift"),
-                group("somegroup4", [
-                    file("somefile7.swift"),
-                ]),
-            ]),
-            group("somegroup1", [
-                file("somefile4.swift"),
-                group("somegroup3", [
-                    file("somefile6.swift"),
-                ]),
-                file("somefile3.swift"),
-            ]),
-        ])
-
-        // When
-        ProjectGroups.sortGroups(group: mainGroup)
-
-        // Then
-        assertGroupsEqual(mainGroup, group("project", [
-            group("somegroup1", [
-                group("somegroup3", [
-                    file("somefile6.swift"),
-                ]),
-                file("somefile3.swift"),
-                file("somefile4.swift"),
-            ]),
-            group("somegroup2", [
-                group("somegroup4", [
-                    file("somefile7.swift"),
-                ]),
-                file("somefile1.swift"),
-                file("somefile2.swift"),
-            ]),
-            file("somefile1.swift"),
-        ]))
-    }
-
-    func test_projectGroupsSort_simpleEmptyFoldersCase() throws {
-        // Given
-        let mainGroup = PBXGroup(children: [
-            file("file3"),
-            file("file1"),
-            file("file4.swift"),
-            file("file2"),
-        ])
-
-        // When
-        ProjectGroups.sortGroups(group: mainGroup)
-
-        // Then
-        assertGroupsEqual(mainGroup, group("project", [
-            file("file1"), // folder references
-            file("file2"),
-            file("file3"),
-            file("file4.swift"),
-        ]))
-    }
-
-    func test_projectGroupsSort_simpleEmptyFoldersAndGroupsCase() throws {
-        // Given
-        let mainGroup = PBXGroup(children: [
-            file("file3"),
-            file("file1"),
-            file("file4.swift"),
-            group("zzzgroup", [
-                file("zz.swift"),
-            ]),
-            file("file2"),
-        ])
-
-        // When
-        ProjectGroups.sortGroups(group: mainGroup)
-
-        // Then
-        assertGroupsEqual(mainGroup, group("project", [
-            group("zzzgroup", [ // groups before files
-                file("zz.swift"),
-            ]),
-            file("file1"), // folder references
-            file("file2"),
-            file("file3"),
-            file("file4.swift"),
-        ]))
-    }
-
-    func test_projectGroupsSort_simpleEmptyFoldersAndGroupsCaseDeeperNesting() throws {
-        // Given
-        let mainGroup = PBXGroup(children: [
-            file("somerootfile.md"),
-            group("rootfolder", [
-                file("file3"),
-                file("file1"),
-                file("file4.swift"),
-                group("zzzgroup", [
-                    file("zz.swift"),
-                    file("aa.swift"),
-                ]),
-                file("file2"),
-            ]),
-        ])
-
-        // When
-        ProjectGroups.sortGroups(group: mainGroup)
-
-        // Then
-        assertGroupsEqual(mainGroup, group("project", [
-            group("rootfolder", [
-                group("zzzgroup", [ // groups before files
-                    file("aa.swift"),
-                    file("zz.swift"),
-                ]),
-                file("file1"), // folder references
-                file("file2"),
-                file("file3"),
-                file("file4.swift"),
-            ]),
-            file("somerootfile.md"),
-        ]))
-    }
-
-    // MARK: - Helpers
-
-    func assertGroupsEqual(_ first: PBXGroup, _ second: PBXGroup, file: StaticString = #file, line: UInt = #line) {
-        XCTAssertEqual(first.flattenedChildren, second.flattenedChildren, file: file, line: line)
-    }
-
-    var references: [PBXFileElement] = []
-
-    func file(_ name: String) -> PBXFileReference {
-        let file = PBXFileReference(path: name)
-        references.append(file)
-        return file
-    }
-
-    func group(_ name: String, _ children: [PBXFileElement]) -> PBXGroup {
-        let group = PBXGroup(children: children, name: name)
-        references.append(group)
-        return group
     }
 }
